examples/apps/model-runner-snowflake-streamlit-demo/requirements.txt
streamlit
landingai
pillow
requests
snowflake-connector-python
snowflake-connector-python[pandas]
